backend/requirements.txt
# --- Core Framework ---
fastapi==0.118.2
uvicorn==0.37.0

# --- Database ---
SQLAlchemy==2.0.30
psycopg2-binary==2.9.6

# --- Environment Config ---
python-dotenv==1.1.1

# Data validation
pydantic==2.7.0
pydantic-settings==2.2.1

# --- External API Calls ---
requests==2.32.5